requirements.txt
prometheus-client~=0.11.0
pyyaml~=6.0
